# internal/docs/swagger.yaml
basePath: /
definitions:
...
  internal_video.likeRequest:
    properties:
      video_id:
        type: integer
    type: object
...
host: 127.0.0.1:8080
info:
  contact: {}
  description: 短视频社区后端 API:账号、视频、Feed、互动、通知。
  title: Video Feed API
  version: "1.0"
paths:
...
  /like/like:
    post:
      consumes:
      - application/json
      description: 通过 RabbitMQ 异步更新计数和热度,失败按通道独立降级到同步写
      parameters:
      - description: 视频 ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_video.likeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 点赞视频
      tags:
      - like
  /like/unlike:
    post:
      consumes:
      - application/json
      parameters:
      - description: 视频 ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_video.likeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 取消点赞
      tags:
      - like
...
securityDefinitions:
  BearerAuth:
    description: Bearer JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"